cleanup: separate some reader code
